### 2023-08-25 #5782 @editorialbot
[PRE REVIEW]: Gibbs Sea Water Oceanographic Toolbox of TEOS-10 implemented in Rust
**Submitting author:** @castelao (<a href="http://orcid.org/0000-0002-6765-0708">Guilherme Castelao</a>)
**Repository:** https://github.com/castelao/GSW-rs
**Branch with paper.md** (empty if default branch): joss
**Version:** 0.1.2
