### 2023-08-26 e0f7ec4
Fixing DOIs, missing and bad syntax
