+++ joss/paper.bib
@@ -11,7 +11,7 @@
   title={TEOS-10/GSW-Python: v3.4.1.post0 (v3.4.1.post0)},
   author={Eric Firing and Filipe and Andrew Barna and Ryan Abernathey},
   year={2021},
-  doi={https://doi.org/10.5281/zenodo.5214122}
+  doi={10.5281/zenodo.5214122}
 }
 
 @article{MILLERO200850,
@@ -22,7 +22,7 @@ number = {1},
 pages = {50-72},
 year = {2008},
 issn = {0967-0637},
-doi = {https://doi.org/10.1016/j.dsr.2007.10.001},
+doi = {10.1016/j.dsr.2007.10.001},
 url = {https://www.sciencedirect.com/science/article/pii/S0967063707002282},
 author = {Frank J. Millero and Rainer Feistel and Daniel G. Wright and Trevor J. McDougall},
 }
@@ -35,7 +35,8 @@ author = {Frank J. Millero and Rainer Feistel and Daniel G. Wright and Trevor J.
   number={1},
   pages={43--114},
   year={2003},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi = {10.1016/S0079-6611(03)00088-0}
 }
 
 @article{feistel2008gibbs,
