### 2023-08-26 #5782 @castelao
Pinging @luizirber (co-author) to follow the discussions here.

Checking DOI corrections with:
