### 2023-08-26 e6d11ae
Initializing documentation for sr_from_sp
