### 2023-08-26 d9b0189
Guidance for missing functions
