### 2023-08-29 dc70c2e
doc: ct_from_pt()
