### 2023-08-29 773e9c6
doc: t90_from_t48()
