### 2023-09-08 15eaed4
typo: License in the README file
