### 2023-09-13 e726f16
doc: Extending documentation
Note that is it still missing some references, but I have to make time
to carefully read each one.
