+++ src/conversions.rs
@@ -10,6 +10,19 @@ use crate::{Error, Result};
 /*
 /// Absolute Salinity Anomaly from Practical Salinity
 ///
+/// # Arguments
+///
+/// * `sp`:
+/// * `p`:
+/// * `lon`:
+/// * `lat`:
+///
+/// # Returns
+///
+/// # References
+///
+/// # Examples
+///
 pub fn deltasa_from_sp(sp: f64, p: f64, lon: f64, lat: f64) -> Result<f64> {
     // Remove out of range values
     if ((p < 100.0) & (sp > 120.0)) | ((p >= 100.0) & (sp > 42.0)) {
@@ -104,13 +117,12 @@ pub fn sr_from_sp(sp: f64) -> f64 {
 ///
 /// If compiled with 'compat', $1/u_{PS}$ is approximated to
 /// 0.995306702338459, which can cause a minimal rounding error.
+///
 /// # References
 ///
 /// Millero, F. J., R. Feistel, D. G. Wright, and T. J. McDougall, 2008: The
 ///   composition of Standard Seawater and the definition of the
 ///   Reference-Composition Salinity Scale, Deep-Sea Res. I, 55, 50-72.
-///
-/// # Examples
 ///
 /// # Examples
 /// ```
@@ -134,6 +146,8 @@ gsw_SA_from_Sstar
 gsw_SP_from_Sstar
 gsw_pt_from_CT
 gsw_t_from_CT
+gsw_pt_from_CT(gsw_ct_from_pt, gsw_gibbs_pt0_pt0)
+gsw_t_from_CT(gsw_ct_from_pt, gsw_pt_from_t)
 */
 
 /// Conservative Temperature from potential temperature
@@ -149,8 +163,11 @@ gsw_t_from_CT
 ///
 /// # Features
 ///
-/// * default: Negative salinity returns Error::NegativeSalinity.
-/// * compat: Negative salinity is assumed to be zero.
+/// * default: Negative salinity returns [Error::NegativeSalinity].
+/// * compat:
+///   * Negative salinity is assumed to be zero.
+///   * S factor is approximated to 0.0248826675584615, which can cause a
+///     minimal error.
 /// * invalidasnan: Negative salinity results in NaN.
 ///
 /// # References
@@ -208,9 +225,6 @@ pub fn ct_from_pt(sa: f64, pt: f64) -> Result<f64> {
 }
 
 /*
-gsw_pt_from_CT(gsw_ct_from_pt, gsw_gibbs_pt0_pt0)
-gsw_t_from_CT(gsw_ct_from_pt, gsw_pt_from_t)
-gsw_CT_from_pt
 gsw_pot_enthalpy_from_pt
 gsw_pt_from_t
 gsw_pt0_from_t(gsw_entropy_part, gsw_entropy_part_zerop, gsw_gibbs_pt0_pt0)
@@ -228,6 +242,9 @@ gsw_t_from_pt0
 /// * `t90`: Temperature ITS-90 \[deg C\]
 ///
 /// # References
+///
+/// Saunders, P. 1990: The International Temperature Scale of 1990, ITS-90.
+///   WOCE Newsletter 10, IOS, Wormley, UK.
 ///
 /// # Examples
 /// ```
@@ -254,6 +271,17 @@ pub fn t90_from_t48(t48: f64) -> f64 {
 /// # Returns
 ///
 /// * `t90`: Temperature ITS-90 \[deg C\]
+///
+/// # Features
+///
+/// * default: Calculate precise $t_{68}$ / 1.00024
+/// * compat: Approximate 1 / 1.00024 to 0.999760057586179, which can cause
+///           a negligible error.
+///
+/// # Notes
+///
[…]
