### 2023-09-13 0daf507
doc: Adding some symbols
