### 2023-09-13 8a461ed
Conversion documentation (#59)
* Initializing documentation for sr_from_sp

* Guidance for missing functions

* Tagging source code in documentation

* doc: ct_from_pt()

* doc: t90_from_t48()

* doc: t68_from_t90()

* Adding Saunders 1990

* Adding Millero 2008 reference

* typo: License in the README file

* refactor: Re-routing t68_from_t90()

* doc: Extending documentation

Note that is it still missing some references, but I have to make time
to carefully read each one.

* doc: Adding some symbols

* Upgrade version to reflect on the public documentation
