+++ src/conversions.rs
@@ -10,6 +10,19 @@ use crate::{Error, Result};
 /*
 /// Absolute Salinity Anomaly from Practical Salinity
 ///
+/// # Arguments
+///
+/// * `sp`:
+/// * `p`:
+/// * `lon`:
+/// * `lat`:
+///
+/// # Returns
+///
+/// # References
+///
+/// # Examples
+///
 pub fn deltasa_from_sp(sp: f64, p: f64, lon: f64, lat: f64) -> Result<f64> {
     // Remove out of range values
     if ((p < 100.0) & (sp > 120.0)) | ((p >= 100.0) & (sp > 42.0)) {
@@ -50,10 +63,30 @@ pub fn deltasa_from_sp(sp: f64, p: f64, lon: f64, lat: f64) -> Result<f64> {
 */
 
 /*
+gsw_deltaSA_from_SP(gsw_sa_from_sp, gsw_sr_from_sp)
 gsw_SA_Sstar_from_SP
 */
 
-/// Reference Salinity from Practical Salinity
+/// Reference Salinity ($S_R$) from Practical Salinity ($S_P$)
+///
+/// # Arguments
+///
+/// * `sp`: Practical salinity (PSS-78) \[unitless\]
+///
+/// # Returns
+///
+/// * `sr`: Reference salinity \[g kg-1\]
+///
+/// # Features
+///
+/// If compiled with 'compat', $u_{PS}$ is approximated to
+/// 1.004715428571429, which can cause a minimal rounding error.
+///
+/// # References
+///
+/// Millero, F. J., R. Feistel, D. G. Wright, and T. J. McDougall, 2008: The
+///   composition of Standard Seawater and the definition of the
+///   Reference-Composition Salinity Scale, Deep-Sea Res. I, 55, 50-72.
 ///
 /// # Examples
 /// ```
@@ -70,7 +103,26 @@ pub fn sr_from_sp(sp: f64) -> f64 {
     }
 }
 
-/// Practical Salinity from Reference Salinity
+/// Practical Salinity ($S_P$) from Reference Salinity ($S_R$)
+///
+/// # Arguments
+///
+/// * `sr`: Reference salinity \[g kg-1\]
+///
+/// # Returns
+///
+/// * `sp`: Practical salinity (PSS-78) \[unitless\]
+///
+/// # Features
+///
+/// If compiled with 'compat', $1/u_{PS}$ is approximated to
+/// 0.995306702338459, which can cause a minimal rounding error.
+///
+/// # References
+///
+/// Millero, F. J., R. Feistel, D. G. Wright, and T. J. McDougall, 2008: The
+///   composition of Standard Seawater and the definition of the
+///   Reference-Composition Salinity Scale, Deep-Sea Res. I, 55, 50-72.
 ///
 /// # Examples
 /// ```
@@ -88,15 +140,37 @@ pub fn sp_from_sr(sr: f64) -> f64 {
 }
 
 /*
-gsw_SP_from_SA
+gsw_SP_from_SA(gsw_sp_from_sa_baltic)
 gsw_Sstar_from_SA
 gsw_SA_from_Sstar
 gsw_SP_from_Sstar
 gsw_pt_from_CT
 gsw_t_from_CT
+gsw_pt_from_CT(gsw_ct_from_pt, gsw_gibbs_pt0_pt0)
+gsw_t_from_CT(gsw_ct_from_pt, gsw_pt_from_t)
 */
 
 /// Conservative Temperature from potential temperature
+///
+/// # Arguments
+///
+/// * `sa`: Absolute salinity \[g kg-1\]
+/// * `pt`: Potential temperature (ITS-90) \[deg C\]
+///
+/// # Returns
+///
+/// * `ct`: Conservative Temperature (ITS-90) \[deg C\]
+///
[…]
