### 2023-09-13 80f5992
add Rust Benchmark (cargo) benchmark result for 8a461edad1dccbdae07b177ce403949e503b5131
