+++ dev/bench/data.js
@@ -1,5 +1,5 @@
 window.BENCHMARK_DATA = {
-  "lastUpdate": 1692893467056,
+  "lastUpdate": 1694584258976,
   "repoUrl": "https://github.com/castelao/GSW-rs",
   "entries": {
     "Rust Benchmark": [
@@ -5998,6 +5998,258 @@ window.BENCHMARK_DATA = {
           {
             "name": "z_from_p",
             "value": 17,
+            "range": "± 0",
+            "unit": "ns/iter"
+          }
+        ]
+      },
+      {
+        "commit": {
+          "author": {
+            "email": "guilherme@castelao.net",
+            "name": "Guilherme Castelão",
+            "username": "castelao"
+          },
+          "committer": {
+            "email": "noreply@github.com",
+            "name": "GitHub",
+            "username": "web-flow"
+          },
+          "distinct": true,
+          "id": "8a461edad1dccbdae07b177ce403949e503b5131",
+          "message": "Conversion documentation (#59)\n\n* Initializing documentation for sr_from_sp\r\n\r\n* Guidance for missing functions\r\n\r\n* Tagging source code in documentation\r\n\r\n* doc: ct_from_pt()\r\n\r\n* doc: t90_from_t48()\r\n\r\n* doc: t68_from_t90()\r\n\r\n* Adding Saunders 1990\r\n\r\n* Adding Millero 2008 reference\r\n\r\n* typo: License in the README file\r\n\r\n* refactor: Re-routing t68_from_t90()\r\n\r\n* doc: Extending documentation\r\n\r\nNote that is it still missing some references, but I have to make time\r\nto carefully read each one.\r\n\r\n* doc: Adding some symbols\r\n\r\n* Upgrade version to reflect on the public documentation",
+          "timestamp": "2023-09-12T23:38:18-06:00",
+          "tree_id": "ede03558f0f247fa359122260be4ca7c420d2338",
+          "url": "https://github.com/castelao/GSW-rs/commit/8a461edad1dccbdae07b177ce403949e503b5131"
+        },
+        "date": 1694584258041,
+        "tool": "cargo",
+        "benches": [
+          {
+            "name": "specvol",
+            "value": 36,
+            "range": "± 0",
+            "unit": "ns/iter"
+          },
+          {
+            "name": "alpha",
+            "value": 64,
+            "range": "± 1",
+            "unit": "ns/iter"
+          },
+          {
+            "name": "beta",
+            "value": 61,
+            "range": "± 0",
+            "unit": "ns/iter"
+          },
+          {
+            "name": "specvol_sso_0",
+            "value": 4,
+            "range": "± 0",
+            "unit": "ns/iter"
+          },
+          {
+            "name": "specvol_anom_standard",
+            "value": 41,
+            "range": "± 0",
+            "unit": "ns/iter"
+          },
+          {
+            "name": "specvol_alpha_beta",
+            "value": 158,
+            "range": "± 0",
+            "unit": "ns/iter"
+          },
+          {
+            "name": "rho",
+            "value": 41,
+            "range": "± 0",
+            "unit": "ns/iter"
+          },
+          {
+            "name": "rho_alpha_beta",
+            "value": 158,
+            "range": "± 0",
+            "unit": "ns/iter"
+          },
+          {
+            "name": "sigma0",
+            "value": 17,
+            "range": "± 0",
+            "unit": "ns/iter"
+          },
+          {
+            "name": "sigma1",
+            "value": 41,
+            "range": "± 0",
+            "unit": "ns/iter"
+          },
+          {
+            "name": "sigma2",
+            "value": 41,
+            "range": "± 0",
+            "unit": "ns/iter"
+          },
+          {
+            "name": "sigma3",
[…]
