### 2023-10-12 #5782 @kthyng
@amadeusine Are you interested in reviewing this submission for the Journal for Open Source Software (JOSS)? I know you may not have much specific domain expertise, but would look to you as focused on reviewing the rust code itself.
Our reviews are done openly in a github issue and are checklist driven — you can get more information here: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#
I would ask that you start your review in the next 3 weeks with the intention of having it finished within 4-6 weeks, given that the process is often iterative with the submitting author(s).

Thanks for your consideration!
