### 2023-11-17 castelao/GSW-rs#61 @efiring
Clarify functionality coverage relative to GSW-C
I think that the readme and the joss paper need to say something about the subset of GSW that is already implemented, and perhaps about how much that subset might (or won't) be expanded.  In particular, comparison to the current state of the C implementation would be useful.  It should be made clear that using the current Rust implementation as a plugin replacement for the C library in the Python implementation will result in considerably reduced functionality.  I haven't looked closely enough to figure out how this is handled in your CI when you do plug in the Rust and then run the Python tests.
