### 2023-11-17 #5988 @efiring
I'm requesting clarification of how the Rust implementation currently compares in coverage to (mainly) GSW-C: https://github.com/castelao/GSW-rs/issues/61
