### 2023-11-17 castelao/GSW-rs#61 @castelao
That's a good point. The goal is to eventually cover everything that GSW-C provides, but we should have it clear what is missing instead of having it implicit by what is not in our documentation. I'll work on that. Thanks,
