### 2023-11-17 2e54ff8
List of functions from GSW-C
Adressing #61.

Let's start with a target list of functions implemented in GSW-C. Maybe
mark what was already implemented in Rust.
