+++ inventory.md
@@ -1,13 +1,92 @@
+
+# Module Convertions
+
+- [X] gsw_sr_from_sp(sp)
+- [X] gsw_ct_from_pt(sa,pt)
+- [X] gsw_z_from_p(p,lat,geo_strf_dyn_height,sea_surface_geopotential)
+- [X] gsw_p_from_z(z,lat, geo_strf_dyn_height, sea_surface_geopotential)
+- [ ] gsw_deltasa_from_sp(sp,p,lon,lat)
+
+
+p_from_abs_pressure
+abs_pressure_from_p
+sp_from_sr
+t68_from_t90
+t90_from_t48
+t90_from_t68
+
+# Module Earth
+
+- [X] gsw_grav(lat,p)
+
+coriollis_parameter
+distance
+
+# Module Practival Salinity
+
+- [X] gsw_sp_from_c(c,t,p)
+- [X] gsw_c_from_sp(sp,t,p)
+- [X] gsw_sp_salinometer(rt,t)
+
+sp_from_r
+r_from_sp
+
+# Module Volume
+
+- [X] gsw_specvol(sa,ct,p)
+- [X] gsw_rho(sa,ct,p)
+- [X] gsw_alpha(sa,ct,p)
+- [X] gsw_beta(sa,ct,p)
+- [X] gsw_alpha_on_beta(sa,ct,p)
+- [X] gsw_specvol_anom_standard(sa,ct,p)
+- [X] gsw_sigma0(sa,ct)
+- [X] gsw_sigma1(sa,ct)
+- [X] gsw_sigma2(sa,ct)
+- [X] gsw_sigma3(sa,ct)
+- [X] gsw_sigma4(sa,ct)
+- [X] gsw_cabbeling(sa,ct,p)
+- [X] gsw_thermobaric(sa,ct,p)
+- [X] gsw_enthalpy(sa,ct,p)
+- [X] gsw_enthalpy_diff (sa, ct, p_shallow, p_deep)
+- [X] gsw_sound_speed(sa,ct,p)
+- [X] gsw_kappa(sa,ct,p)
+- [X] gsw_internal_energy(sa,ct,p)
+- [X] gsw_dynamic_enthalpy(sa,ct,p)
+- [X] gsw_sa_from_rho(rho,ct,p)
+- [X] gsw_ct_maxdensity (sa, p)
+- [ ] gsw_specvol_ice (t, p)
+
+
+rho_alpha_beta
+rho_first_derivatives
+rho_second_derivatives
+specvol_alpha_beta
+specvol_first_derivatives
+specvol_second_derivatives
+specvol_first_derivatives_wrt_enthalpy
+specvol_second_derivatives_wrt_enthalpy
+specvol_anom
+internal_energy_first_derivatives
+internal_energy_second_derivatives
+enthalpy_first_derivatives
+enthalpy_second_derivatives
+
+# Internal Functions
+
+- [X] gsw_specvol_sso_0(p)
+- [X] gsw_enthalpy_sso_0(p)
+- [X] gsw_hill_ratio_at_sp2(t)
+- [X] gsw_gibbs(ns,nt,np,sa,t,p)
+- [X] gsw_gibbs_ice (nt, np, t, p)
+
+
+# Missing in Rust implementation
+
 - [ ] gsw_adiabatic_lapse_rate_from_ct(sa,ct,p)
 - [ ] gsw_adiabatic_lapse_rate_ice (t, p)
[…]
