### 2023-11-19 62c5ad6
Listing missing functions in module conversions
Clearly conversions should be the first target.
