+++ inventory.md
@@ -1,16 +1,32 @@
 
 # Module Convertions
 
+- [ ] gsw_deltasa_from_sp(sp,p,lon,lat)
 - [X] gsw_sr_from_sp(sp)
+- [X] gsw_sp_from_sr(sr)
+- [ ] gsw_sp_from_sa(sa,p,lon,lat)
+- [ ] gsw_sstar_from_sa(sa,p,lon,lat)
+- [ ] gsw_sa_from_sstar(sstar,p,lon,lat)
+- [ ] gsw_sp_from_sstar(sstar,p,lon,lat)
+- [ ] gsw_pt_from_ct(sa,ct)
+- [ ] gsw_t_from_ct(sa,ct,p)
 - [X] gsw_ct_from_pt(sa,pt)
+- [ ] gsw_pt_from_t(sa,t,p,p_ref)
+- [ ] gsw_pt0_from_t(sa,t,p)
 - [X] gsw_z_from_p(p,lat,geo_strf_dyn_height,sea_surface_geopotential)
 - [X] gsw_p_from_z(z,lat, geo_strf_dyn_height, sea_surface_geopotential)
-- [ ] gsw_deltasa_from_sp(sp,p,lon,lat)
+- [ ] gsw_entropy_from_ct (sa, ct)
+- [ ] gsw_ct_from_entropy (sa, entropy)
+- [ ] gsw_entropy_from_pt (sa, pt)
+- [ ] gsw_pt_from_entropy (sa, entropy)
+- [ ] gsw_adiabatic_lapse_rate_from_ct(sa,ct,p)
+
+
+
 
 
 p_from_abs_pressure
 abs_pressure_from_p
-sp_from_sr
 t68_from_t90
 t90_from_t48
 t90_from_t68
@@ -27,9 +43,26 @@ distance
 - [X] gsw_sp_from_c(c,t,p)
 - [X] gsw_c_from_sp(sp,t,p)
 - [X] gsw_sp_salinometer(rt,t)
+- [ ] gsw_sp_from_sk(sk)
 
-sp_from_r
-r_from_sp
+## Missing on GSW-C
+- sp_from_r
+- r_from_sp
+
+## Not implemented yet
+- sp_from_sk
+
+# Module Spiceness
+
+- [ ] gsw_spiciness0(sa,ct)
+- [ ] gsw_spiciness1(sa,ct)
+- [ ] gsw_spiciness2(sa,ct)
+
+## Missing on GSW-C
+
+- [ ] gsw_SA_CT_from_sigma0_spiciness0
+- [ ] gsw_SA_CT_from_sigma1_spiciness1
+- [ ] gsw_SA_CT_from_sigma2_spiciness2
 
 # Module Volume
 
@@ -71,6 +104,21 @@ internal_energy_second_derivatives
 enthalpy_first_derivatives
 enthalpy_second_derivatives
 
+# Module geostrophic streamfunctions, acoustic travel time and geostrophic velocity
+- [ ] gsw_geo_strf_dyn_height (sa, ct, p, p_ref)
+
+# Module Volume Exact
+
+# Module Derivatives
+
+# Module Ice
+
+# Module vertical stability and interpolation
+
+# Module thermodynamic interaction between ice and seawater
+
+# Module Absolute Salinity (SA), Preformed Salinity (Sstar) and Conservative Temperature (CT)
+
 # Internal Functions
 
 - [X] gsw_specvol_sso_0(p)
[…]
