### 2023-11-19 449717b
doc: Some guidance on how to use it (#60)
Anyone new to Rust might feel lost how to start. Let's give a minumum
guidance here.
