### 2023-11-19 4a70ecb
add Rust Benchmark (cargo) benchmark result for 449717b5be4530bdda2047d31de3e4bf8f6dd07c
