### 2023-11-20 d6aab4f
Minimum explanation on what is the inventory
