### 2023-11-27 castelao/GSW-rs#63 @wentasah
README could contain more information about the project
Why rewrite in rust? What are the benefits compared to the C implementation? If you are faster, mention that and write how much. If you are “safer”, mention why it matters.

I see that this is answered in `doc/talks`, but this document is not easily accessible (one has to download the repo and open the file locally). It would be great if the main goals are mentioned directly in the top-level `README.md`.
