+++ src/lib.rs
@@ -3,7 +3,7 @@
 //! Unofficial Gibbs Sea Water Oceanographic Toolbox of TEOS-10,
 //! based on version 3.06.12, implemented in Rust.
 //!
-//! Note that we do follow TEOS-10 manual and references, but his library is
+//! Note that we do follow TEOS-10 manual and references, but this library is
 //! not endorsed by TEOS-10 committee. For the official TEOS-10 documentation
 //! and other software implementations, check: <http://www.teos-10.org>
 //!
