### 2023-11-27 9ffc794
add Rust Benchmark (cargo) benchmark result for 7cea3420d7d35d0d6f11b083cb480b8607e0ddeb
