### 2023-12-01 4039cac
Updating readme about missing functions
