### 2023-12-01 7062d11
Adding the 'gsw_' prefix
Although we don't use it in Rust, since it is explicit by the crate plus
module (ex. gsw::earth::distance), let's add it here for consistency.
