+++ inventory.md
@@ -1,0 +1,238 @@
+
+Currently, GSW-Rust do not implement all functions already available in
+GSW-C or GSW-Matlab. This document gives an idea of what is missing. If
+interested in some specific function, please check our official manual as
+the best reference for what we have already done.
+
+We organized the functions in modules following the toolbox card from
+TEOS-10 main document. The functions unchecked are missing in GSW-Rust.
+The goal is to implement all functions described in TEOS-10, but the
+priority is to do it right.
+
+
+# Conversions
+
+- [ ] gsw_deltasa_from_sp(sp,p,lon,lat)
+- [ ] gsw_sstar_from_sp(sp,p,lon,lat)
+- [X] gsw_sr_from_sp(sp)
+- [X] gsw_sp_from_sr(sr)
+- [ ] gsw_sp_from_sa(sa,p,lon,lat)
+- [ ] gsw_sstar_from_sa(sa,p,lon,lat)
+- [ ] gsw_sa_from_sstar(sstar,p,lon,lat)
+- [ ] gsw_sp_from_sstar(sstar,p,lon,lat)
+- [ ] gsw_pt_from_ct(sa,ct)
+- [ ] gsw_t_from_ct(sa,ct,p)
+- [X] gsw_ct_from_pt(sa,pt)
+- [ ] gsw_pt_from_t(sa,t,p,p_ref)
+- [ ] gsw_pt0_from_t(sa,t,p)
+- [X] gsw_z_from_p(p,lat,geo_strf_dyn_height,sea_surface_geopotential)
+- [X] gsw_p_from_z(z,lat, geo_strf_dyn_height, sea_surface_geopotential)
+- [ ] gsw_entropy_from_ct (sa, ct)
+- [ ] gsw_ct_from_entropy (sa, entropy)
+- [ ] gsw_entropy_from_pt (sa, pt)
+- [ ] gsw_pt_from_entropy (sa, entropy)
+- [ ] gsw_entropy_from_t(sa,t,p)
+- [ ] gsw_adiabatic_lapse_rate_from_ct(sa,ct,p)
+
+
+
+## Missing on GSW-C
+
+- [ ] gsw_pot_enthalpy_from_pt
+- [ ] gsw_t_from_pt0
+- [X] gsw_t90_from_t48()
+- [X] gsw_t90_from_t68()
+- [ ] gsw_z_from_depth
+- [ ] gsw_depth_from_z
+- [X] gsw_abs_pressure_from_p()
+- [X] gsw_p_from_abs_pressure()
+- [ ] gsw_t_from_entropy()
+- [ ] gsw_adiabatic_lapse_rate_from_t()
+- [ ] gsw_molality_from_sa
+- [ ] gsw_ionic_strength_from_sa
+- [X] gsw_t68_from_t90()
+
+# Earth
+
+- [X] gsw_grav(lat,p)
+
+## Missing on GSW-C
+
+- [X] coriollis_parameter()
+- [X] gsw_distance()
+
+# Practical Salinity
+
+- [X] gsw_sp_from_c(c,t,p)
+- [X] gsw_c_from_sp(sp,t,p)
+- [X] gsw_sp_salinometer(rt,t)
+- [ ] gsw_sp_from_sk(sk)
+
+## Missing on GSW-C
+
+- [X] gsw_sp_from_r
+- [X] gsw_r_from_sp
+
+# Spiceness
+
[…]
