### 2023-12-01 28f017c
add Rust Benchmark (cargo) benchmark result for 82c0f9757c54ac09d2301583fba15becbb3c48b5
