+++ dev/bench/data.js
@@ -1,5 +1,5 @@
 window.BENCHMARK_DATA = {
-  "lastUpdate": 1701104559441,
+  "lastUpdate": 1701409504481,
   "repoUrl": "https://github.com/castelao/GSW-rs",
   "entries": {
     "Rust Benchmark": [
@@ -6748,6 +6748,258 @@ window.BENCHMARK_DATA = {
           {
             "name": "sp_salinometer",
             "value": 8,
+            "range": "± 0",
+            "unit": "ns/iter"
+          },
+          {
+            "name": "z_from_p",
+            "value": 13,
+            "range": "± 0",
+            "unit": "ns/iter"
+          }
+        ]
+      },
+      {
+        "commit": {
+          "author": {
+            "email": "guilherme@castelao.net",
+            "name": "Guilherme Castelão",
+            "username": "castelao"
+          },
+          "committer": {
+            "email": "noreply@github.com",
+            "name": "GitHub",
+            "username": "web-flow"
+          },
+          "distinct": true,
+          "id": "82c0f9757c54ac09d2301583fba15becbb3c48b5",
+          "message": "List of functions from GSW-C (#62)\n\n* List of functions from GSW-C\r\n\r\nAdressing #61.\r\n\r\nLet's start with a target list of functions implemented in GSW-C. Maybe\r\nmark what was already implemented in Rust.\r\n\r\n* Using checkboxes\r\n\r\n* Updating already implemented list\r\n\r\n* Listing missing functions in module conversions\r\n\r\nClearly conversions should be the first target.\r\n\r\n* Minimum explanation on what is the inventory\r\n\r\n* Organizing this list\r\n\r\n* typo: Thanks @efiring\r\n\r\n* Re-organizing lists per modules\r\n\r\n* Updating readme about missing functions\r\n\r\n* Adding the 'gsw_' prefix\r\n\r\nAlthough we don't use it in Rust, since it is explicit by the crate plus\r\nmodule (ex. gsw::earth::distance), let's add it here for consistency.",
+          "timestamp": "2023-11-30T22:34:45-07:00",
+          "tree_id": "921271da516d47eebcfa1057a111e608924f1939",
+          "url": "https://github.com/castelao/GSW-rs/commit/82c0f9757c54ac09d2301583fba15becbb3c48b5"
+        },
+        "date": 1701409503540,
+        "tool": "cargo",
+        "benches": [
+          {
+            "name": "specvol",
+            "value": 27,
+            "range": "± 0",
+            "unit": "ns/iter"
+          },
+          {
+            "name": "alpha",
+            "value": 46,
+            "range": "± 0",
+            "unit": "ns/iter"
+          },
+          {
+            "name": "beta",
+            "value": 47,
+            "range": "± 0",
+            "unit": "ns/iter"
+          },
+          {
+            "name": "specvol_sso_0",
+            "value": 3,
+            "range": "± 0",
+            "unit": "ns/iter"
+          },
+          {
+            "name": "specvol_anom_standard",
+            "value": 32,
+            "range": "± 0",
+            "unit": "ns/iter"
+          },
+          {
+            "name": "specvol_alpha_beta",
+            "value": 136,
+            "range": "± 0",
+            "unit": "ns/iter"
+          },
+          {
+            "name": "rho",
+            "value": 33,
+            "range": "± 0",
+            "unit": "ns/iter"
+          },
+          {
+            "name": "rho_alpha_beta",
+            "value": 124,
+            "range": "± 0",
+            "unit": "ns/iter"
+          },
+          {
+            "name": "sigma0",
+            "value": 17,
+            "range": "± 0",
+            "unit": "ns/iter"
+          },
+          {
+            "name": "sigma1",
[…]
