### 2023-12-01 27c44ec
add Rust Benchmark (cargo) benchmark result for 92be3c9b9cfee6b9d56d022e95ee2eced3eec5bd
