### 2023-12-01 castelao/GSW-rs#63 @castelao
A reminder for myself - Add a statement of need in the documentation. I have no opinion yet if explicit or implicit.
