### 2023-12-02 #5988 @efiring
Minor suggestions:

Line 37, "we propose" -> "we present"
Lines 40-42, the last sentence of the paper: it doesn't quite make sense to me.  The simplest solution would be to delete it.
References: in the last one, the UNESCO manual, the parts in parentheses and brackets don't seem to belong.  I would expect the citation form here to match what is in the README, and whatever is provided by the TEOS-10 web site.
