### 2023-12-02 d93e910
Fixing reference
Type techreport might be more adequate. Other minor fixes.
