+++ joss/paper.bib
@@ -1,10 +1,11 @@
-@article{GSW-m,
-  title={Getting started with TEOS-10 and the Gibbs Seawater (GSW) oceanographic toolbox},
+@techreport{GSW-m,
+  title={Getting started with {TEOS-10} and the {G}ibbs {S}eawater ({GSW}) oceanographic toolbox},
   author={McDougall, Trevor J and Barker, Paul M},
-  journal={{SCOR}/{IAPSO}},
-  volume={WG 127},
-  pages={1--28},
-  year={2011}
+  publisher={{SCOR}/{IAPSO}},
+  series={WG 127},
+  pages={28pp.},
+  year={2011},
+  note={ISBN 978-0-646-55621-5}
 }
 
 @article{GSW-Python,
