### 2023-12-02 31beff0
Editing confusing sentence
@efiring, thanks for the note. Is it bette now? The point behind is that
with `features='compat'` we guarantee compatibility with GSW-m. While
GSW-C has some minor differences.
