+++ joss/paper.md
@@ -65,8 +65,7 @@ some features and conveniences of modern languages. Here we present an
 alternative using Rust language, resulting in comparable performance to GSW-C,
 while providing an efficient framework that accelerates the development effort
 and minimizes errors. For embedded systems, GSW-rs is a requirement for a
-Rust firmware able to make sense of sensor measurements in real time. Beyond
-embedded, GSW-rs allows for the reproduction of the GSW-m, thus guaranteeing
-reproducibility for the Open Source community.
+pure Rust firmware able to make sense of sensor measurements in real time.
+Beyond embedded applications, GSW-rs is also compatible with regular computers, thus providing an open-source alternative for the scientific community that guarantees reproducibility with GSW-m.
 
 # References
