### 2023-12-02 7e0cec6
Bug, manual type is expected to be a software
Somehow JOSS processing assumes that a bib reference type manual should
be from a software and adds '[Computer software]' to the refence.

Let's see if a techreport resolves that.
