+++ joss/paper.bib
@@ -51,7 +51,7 @@ author = {Frank J. Millero and Rainer Feistel and Daniel G. Wright and Trevor J.
   publisher={Elsevier}
 }
 
-@manual{TEOS-10,
+@techreport{TEOS-10,
   title={The international thermodynamic equation of seawater – 2010: Calculation and use of thermodynamic properties},
   edition={June 2015},
   volume={Intergovernmental Oceanographic Commission, Manuals and Guides No. 56},
