+++ joss/paper.bib
@@ -53,12 +53,12 @@ author = {Frank J. Millero and Rainer Feistel and Daniel G. Wright and Trevor J.
 
 @techreport{TEOS-10,
   title={The international thermodynamic equation of seawater – 2010: Calculation and use of thermodynamic properties},
+  author={{IOC, SCOR, and IAPSO}},
   edition={June 2015},
-  volume={Intergovernmental Oceanographic Commission, Manuals and Guides No. 56},
+  series={Intergovernmental Oceanographic Commission, Manuals and Guides No. 56},
   year={2010},
   publisher={UNESCO},
-  note={IOC/2010/MG/56 Rev.},
-  key={{IOC, SCOR, and IAPSO}, 2010}
+  note={IOC/2010/MG/56 Rev.}
 }
 
 @article{GSW-Julia,
