+++ joss/paper.bib
@@ -3,7 +3,7 @@
   author={McDougall, Trevor J and Barker, Paul M},
   publisher={{SCOR}/{IAPSO}},
   series={WG 127},
-  pages={28pp.},
+  pages={1--28},
   year={2011},
   note={ISBN 978-0-646-55621-5}
 }
