### 2023-12-02 3f67aef
Adding missing total pages
It's up to JOSS references style to decide showing it or not.
