+++ joss/paper.bib
@@ -58,7 +58,7 @@ author = {Frank J. Millero and Rainer Feistel and Daniel G. Wright and Trevor J.
   series={Intergovernmental Oceanographic Commission, Manuals and Guides No. 56},
   year={2010},
   publisher={UNESCO},
-  note={IOC/2010/MG/56 Rev.}
+  note={196 pp.}
 }
 
 @article{GSW-Julia,
