### 2023-12-02 19f310c
Improving compatibility comment and not on limited coverage
Perfect! Clear message. Thanks @efiring!
