### 2023-12-06 374c34d
Using manual instead of article
For the software without DOI, reference type 'manual' might be better
than 'article'.
