+++ joss/paper.bib
@@ -61,7 +61,7 @@ author = {Frank J. Millero and Rainer Feistel and Daniel G. Wright and Trevor J.
   note={196 pp.}
 }
 
-@article{GSW-Julia,
+@manual{GSW-Julia,
   title={TEOS-10/GibbsSeaWater.jl},
   author={Alexander Barth and et. al.},
   year={2020},
@@ -69,7 +69,7 @@ author = {Frank J. Millero and Rainer Feistel and Daniel G. Wright and Trevor J.
   note={GitHub Repository.}
 }
 
-@article{GSW-R,
+@manual{GSW-R,
   title={TEOS-10/GSW-R},
   author={Dan Kelley and Clark Richards},
   year={2022},
@@ -77,7 +77,7 @@ author = {Frank J. Millero and Rainer Feistel and Daniel G. Wright and Trevor J.
   note={GitHub Repository.}
 }
 
-@article{GSW-FORTRAN,
+@manual{GSW-FORTRAN,
   title={TEOS-10/GSW-FORTRAN},
   author={David Jackett and Paul Barker and Glenn Hyland},
   year={2017},
@@ -85,7 +85,7 @@ author = {Frank J. Millero and Rainer Feistel and Daniel G. Wright and Trevor J.
   note={GitHub Repository.}
 }
 
-@article{GSW-C,
+@manual{GSW-C,
   title={},
   author={Frank Delahoyde and Eric Firing and et. al.},
   year={2022},
