### 2023-12-06 a396aa2
Unlock libm version to 0.2
