### 2023-12-06 e20bcdc
Adding categories to Cargo.toml
