+++ Cargo.toml
@@ -32,7 +32,7 @@ regex = "1.5.5"
 
 
 [build-dependencies]
-cbindgen = { version = "0.18.0", optional = true }
+cbindgen = { version = "0.26.0", optional = true }
 
 [features]
 default = []
