### 2023-12-06 5db4777
Adding metadata 'rust-version'
