+++ Cargo.toml
@@ -9,6 +9,7 @@ license = "BSD-3-Clause"
 repository = "https://github.com/castelao/GSW-rs"
 keywords = ["oceanography", "seawater"]
 categories = ["science", "no-std"]
+rust-version = "1.64.0"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
