### 2023-12-06 a269bf4
Lowering rust-version to 1.59
This is the latest version that nix-portable can handle, but it would be
nice sometime soon to bump to 1.68
