### 2023-12-06 b6db58b
Updating crate (#67)
* Unlock libm version to 0.2

* Adding categories to Cargo.toml

* Updating cbindgen

* Adding metadata 'rust-version'

* Updating Cargo.lock

* libm at least 0.2.8

* Updating Cargo.lock - libm

* Let's try a newwer nix-portable

* Lowering rust-version to 1.59

This is the latest version that nix-portable can handle, but it would be
nice sometime soon to bump to 1.68
