+++ Cargo.toml
@@ -8,6 +8,8 @@ documentation = "https://docs.rs/gsw"
 license = "BSD-3-Clause"
 repository = "https://github.com/castelao/GSW-rs"
 keywords = ["oceanography", "seawater"]
+categories = ["science", "no-std"]
+rust-version = "1.59.0"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
@@ -15,7 +17,7 @@ keywords = ["oceanography", "seawater"]
 members = ["convert_refdata"]
 
 [dependencies]
-libm = "=0.2.3"
+libm = "0.2.8"
 libc = { version = "0.2", optional = true }
 thiserror = { version = "1.0.24", optional = true }
 num-complex = "0.4.3"
@@ -31,7 +33,7 @@ regex = "1.5.5"
 
 
 [build-dependencies]
-cbindgen = { version = "0.18.0", optional = true }
+cbindgen = { version = "0.26.0", optional = true }
 
 [features]
 default = []
