### 2023-12-06 0089bba
add Rust Benchmark (cargo) benchmark result for b6db58bd123fc359e14415fddf8eab5adf13cf5e
