+++ dev/bench/data.js
@@ -1,5 +1,5 @@
 window.BENCHMARK_DATA = {
-  "lastUpdate": 1701409886637,
+  "lastUpdate": 1701874926047,
   "repoUrl": "https://github.com/castelao/GSW-rs",
   "entries": {
     "Rust Benchmark": [
@@ -7252,6 +7252,258 @@ window.BENCHMARK_DATA = {
           {
             "name": "sp_salinometer",
             "value": 9,
+            "range": "± 0",
+            "unit": "ns/iter"
+          },
+          {
+            "name": "z_from_p",
+            "value": 13,
+            "range": "± 0",
+            "unit": "ns/iter"
+          }
+        ]
+      },
+      {
+        "commit": {
+          "author": {
+            "email": "guilherme@castelao.net",
+            "name": "Guilherme Castelão",
+            "username": "castelao"
+          },
+          "committer": {
+            "email": "noreply@github.com",
+            "name": "GitHub",
+            "username": "web-flow"
+          },
+          "distinct": true,
+          "id": "b6db58bd123fc359e14415fddf8eab5adf13cf5e",
+          "message": "Updating crate (#67)\n\n* Unlock libm version to 0.2\r\n\r\n* Adding categories to Cargo.toml\r\n\r\n* Updating cbindgen\r\n\r\n* Adding metadata 'rust-version'\r\n\r\n* Updating Cargo.lock\r\n\r\n* libm at least 0.2.8\r\n\r\n* Updating Cargo.lock - libm\r\n\r\n* Let's try a newwer nix-portable\r\n\r\n* Lowering rust-version to 1.59\r\n\r\nThis is the latest version that nix-portable can handle, but it would be\r\nnice sometime soon to bump to 1.68",
+          "timestamp": "2023-12-06T07:51:44-07:00",
+          "tree_id": "d945a92582b1696d2896e1b35de8aecaeb1b45a2",
+          "url": "https://github.com/castelao/GSW-rs/commit/b6db58bd123fc359e14415fddf8eab5adf13cf5e"
+        },
+        "date": 1701874925467,
+        "tool": "cargo",
+        "benches": [
+          {
+            "name": "specvol",
+            "value": 27,
+            "range": "± 0",
+            "unit": "ns/iter"
+          },
+          {
+            "name": "alpha",
+            "value": 45,
+            "range": "± 0",
+            "unit": "ns/iter"
+          },
+          {
+            "name": "beta",
+            "value": 52,
+            "range": "± 0",
+            "unit": "ns/iter"
+          },
+          {
+            "name": "specvol_sso_0",
+            "value": 3,
+            "range": "± 0",
+            "unit": "ns/iter"
+          },
+          {
+            "name": "specvol_anom_standard",
+            "value": 32,
+            "range": "± 0",
+            "unit": "ns/iter"
+          },
+          {
+            "name": "specvol_alpha_beta",
+            "value": 127,
+            "range": "± 0",
+            "unit": "ns/iter"
+          },
+          {
+            "name": "rho",
+            "value": 33,
+            "range": "± 0",
+            "unit": "ns/iter"
+          },
+          {
+            "name": "rho_alpha_beta",
+            "value": 137,
+            "range": "± 0",
+            "unit": "ns/iter"
+          },
+          {
+            "name": "sigma0",
+            "value": 17,
+            "range": "± 1",
+            "unit": "ns/iter"
+          },
+          {
+            "name": "sigma1",
+            "value": 34,
+            "range": "± 0",
+            "unit": "ns/iter"
+          },
+          {
+            "name": "sigma2",
+            "value": 34,
+            "range": "± 0",
+            "unit": "ns/iter"
+          },
+          {
+            "name": "sigma3",
[…]
