### 2023-12-16 #5988 @wentasah
Yes, I'm waiting for https://github.com/castelao/GSW-rs/issues/63. Strictly speaking, I could check the boxes now, because the "missing" information is actually contained in the repo, but it's a bit hidden.

@castelao I think that adding two or three sentences to the readme, perhaps referring the reader to the slides for details, would be sufficient.
