### 2023-12-16 3f5a458
Brief explanation why in Rust
In response to #63.
