### 2023-12-16 castelao/GSW-rs#63 @castelao
@wentasah , thanks again for raising this issue. Does #68 covers everything? You might also enjoy this [benchmark](https://gsw.castelao.net/dev/bench/).
