### 2023-12-16 #5988 @wentasah
The mentioned issue was addressed by a PR so I checked the remaining check boxes.
