### 2023-12-16 184b2ed
Why re-write it in Rust (#68)
* Brief explanation why in Rust

In response to #63.

* Updating link to Seas' talk
