### 2023-12-16 9c912d2
add Rust Benchmark (cargo) benchmark result for 184b2ed0f10402eb3bf2fc3312d90d4b95c543cc
