### 2023-12-19 6587447
JOSS (#45)
* Adding core references

* Creating basic information on paper

* Paper first version

* Compile paper draft on every update

* Fixing references, spaces instead of tabs

* Typo in the TEOS-10 bib alias

* Fixing path for the PDF output

* Missing reference to GSW-Python

It seems like it was the only implementation with a doi missing. It
might be worth figure out how to cite the other implementation, even if
those other ones don't have a doi or any proper citation descriptor.

* Adding references to other repositories

* Fixing DOIs, missing and bad syntax

* Fixing reference

Typ
[…]
