+++ joss/paper.bib
@@ -1,0 +1,94 @@
+@techreport{GSW-m,
+  title={Getting started with {TEOS-10} and the {G}ibbs {S}eawater ({GSW}) oceanographic toolbox},
+  author={McDougall, Trevor J and Barker, Paul M},
+  publisher={{SCOR}/{IAPSO}},
+  series={WG 127},
+  pages={1--28},
+  year={2011},
+  note={ISBN 978-0-646-55621-5}
+}
+
+@article{GSW-Python,
+  title={TEOS-10/GSW-Python: v3.4.1.post0 (v3.4.1.post0)},
+  author={Eric Firing and Filipe and Andrew Barna and Ryan Abernathey},
+  year={2021},
+  doi={10.5281/zenodo.5214122}
+}
+
+@article{MILLERO200850,
+title = {The composition of Standard Seawater and the definition of the Reference-Composition Salinity Scale},
+journal = {Deep Sea Research Part I: Oceanographic Research Papers},
+volume = {55},
+number = {1},
+pages = {50-72},
+year = {2008},
+issn = {0967-0637},
+doi = {10.1016/j.dsr.2007.10.001},
+url = {https://www.sciencedirect.com/science/article/pii/S0967063707002282},
+author = {Frank J. Millero and Rainer Feistel and Daniel G. Wright and Trevor J. McDougall},
+}
+
+@article{feistel2003new,
+  title={A new extended Gibbs thermodynamic potential of seawater},
+  author={Feistel, Rainer},
+  journal={Progress in Oceanography},
+  volume={58},
+  number={1},
+  pages={43--114},
+  year={2003},
+  publisher={Elsevier},
+  doi = {10.1016/S0079-6611(03)00088-0}
+}
+
+@article{feistel2008gibbs,
+  title={A Gibbs function for seawater thermodynamics for- 6 to 80 C and salinity up to 120 g kg--1},
+  author={Feistel, Rainer},
+  journal={Deep Sea Research Part I: Oceanographic Research Papers},
+  volume={55},
+  number={12},
+  pages={1639--1671},
+  year={2008},
+  publisher={Elsevier}
+}
+
+@techreport{TEOS-10,
+  title={The international thermodynamic equation of seawater – 2010: Calculation and use of thermodynamic properties},
+  author={{IOC, SCOR, and IAPSO}},
+  edition={June 2015},
+  series={Intergovernmental Oceanographic Commission, Manuals and Guides No. 56},
+  year={2010},
+  publisher={UNESCO},
+  note={196 pp.}
+}
+
+@manual{GSW-Julia,
+  title={TEOS-10/GibbsSeaWater.jl},
+  author={Alexander Barth and et. al.},
+  year={2020},
+  url={https://github.com/TEOS-10/GibbsSeaWater.jl},
+  note={GitHub Repository.}
+}
+
+@manual{GSW-R,
+  title={TEOS-10/GSW-R},
+  author={Dan Kelley and Clark Richards},
+  year={2022},
+  url={https://github.com/TEOS-10/GSW-R},
+  note={GitHub Repository.}
+}
+
+@manual{GSW-FORTRAN,
+  title={TEOS-10/GSW-FORTRAN},
+  author={David Jackett and Paul Barker and Glenn Hyland},
+  year={2017},
+  url={https://github.com/TEOS-10/GSW-Fortran},
+  note={GitHub Repository.}
+}
+
+@manual{GSW-C,
+  title={},
+  author={Frank Delahoyde and Eric Firing and et. al.},
+  year={2022},
+  url={https://github.com/TEOS-10/GSW-C},
+  note={GitHub Repository.}
+}
