### 2023-12-19 8fff6ed
add Rust Benchmark (cargo) benchmark result for 658744724e1e144a9e45c2708b56b9627b31e3da
