+++ README.md
@@ -25,7 +25,7 @@ We presented about goals and progress (as of Apr 2022) at the
 
 ## Minimum supported Rust version
 
-Currently the minimum supported Rust version is 1.64.0
+Currently the minimum supported Rust version is 1.68.0
 
 ## Installing
 
