+++ Cargo.toml
@@ -28,7 +28,7 @@ num-complex = "0.4.3"
 [dev-dependencies]
 heapless = "0.7"
 inline-c = "0.1"
-postcard = "0.7.0"
+postcard = "1.0.2"
 serde = "1.0.126"
 criterion = "0.3.6"
 # Indirect dependency that address CVE-2022-24713
