+++ Cargo.toml
@@ -31,9 +31,6 @@ inline-c = "0.1"
 postcard = "1.0.2"
 serde = "1.0.126"
 criterion = "0.3.6"
-# Indirect dependency that address CVE-2022-24713
-regex = "1.5.5"
-
 
 [build-dependencies]
 cbindgen = { version = "0.26.0", optional = true }
