### 2023-12-21 1ba9129
Bump GSW to 0.2
