+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "gsw"
-version = "0.1.4"
+version = "0.2.0"
 authors = [
   "Guilherme Castelão <guilherme@castelao.net>",
   "Luiz Irber <luiz.irber@gmail.com>"
