### 2023-12-21 03baf48
Bump minimum version (#69)
* Minimum release 1.68 to match Ferrocene

* style: Breaking line of authors

* Bump postcard to 1.0

* Bumping regex version

* Bump GSW to 0.2
