+++ .github/workflows/ci.yml
@@ -265,11 +265,11 @@ jobs:
       - uses: actions/checkout@v2
 
       - name: check if README matches MSRV defined here
-        run: grep '1.64.0' README.md
+        run: grep '1.68.0' README.md
 
       - uses: actions-rs/toolchain@v1
         with:
-          toolchain: 1.64.0
+          toolchain: 1.68.0
           override: true
 
       - name: Check if it builds
