+++ Cargo.toml
@@ -1,7 +1,10 @@
 [package]
 name = "gsw"
-version = "0.1.4"
-authors = ["Guilherme Castelão <guilherme@castelao.net>", "Luiz Irber <luiz.irber@gmail.com>"]
+version = "0.2.0"
+authors = [
+  "Guilherme Castelão <guilherme@castelao.net>",
+  "Luiz Irber <luiz.irber@gmail.com>"
+  ]
 edition = "2018"
 description = "TEOS-10 v3.06.12 Gibbs Seawater Oceanographic Toolbox in Rust"
 documentation = "https://docs.rs/gsw"
@@ -25,12 +28,9 @@ num-complex = "0.4.3"
 [dev-dependencies]
 heapless = "0.7"
 inline-c = "0.1"
-postcard = "0.7.0"
+postcard = "1.0.2"
 serde = "1.0.126"
 criterion = "0.3.6"
-# Indirect dependency that address CVE-2022-24713
-regex = "1.5.5"
-
 
 [build-dependencies]
 cbindgen = { version = "0.26.0", optional = true }
