### 2023-12-21 3836d67
add Rust Benchmark (cargo) benchmark result for 03baf488228d1d928cd03b72364fd5c397323adb
