### 2023-12-22 024ce88
add Rust Benchmark (cargo) benchmark result for 74cb483cfd1d26e91388af67b09561f731277ace
