### 2023-12-22 24b33d6
Version 0.2.1 - JOSS publication (#71)
