### 2023-12-22 c86eafe
add Rust Benchmark (cargo) benchmark result for 24b33d6983656921e910e20712a00dbecd461186
