### 2023-12-22 #5988 @castelao
@kthyng, I'm not able to checkmark the latest list, but my items are done. Let's use the 0.2.0 release with the Zenodo's archiving below. Thank you so much!

Latest release: 0.2.0

DOI: [10.5281/zenodo.10422483](https://doi.org/10.5281/zenodo.10422483)
