### 2024-01-02 #5988 @kthyng
@castelao I see you have your version selected but I'll just suggest that this could be a good time for declaring a nice new version like v1.0.0.
