### 2024-01-02 #5988 @kthyng
@castelao I should have checked earlier but failed to, but now I see that your license isn't OSI approved. Can you change licenses?
