### 2024-01-03 #5988 @castelao
@kthyng , thanks for your suggestion. I would rather use v1.0.0 when I'm fully compatible with the GSW-C.
