### 2024-01-03 fce32f2
add Rust Benchmark (cargo) benchmark result for 289c10bcda527122cedc338553f31c960309412f
